change number to standard u32

+++ src/main.rs
@@ -3,12 +3,12 @@ extern crate crossbeam;
 use std::env;
 
 pub struct Collatz {
-    curr: u128,
+    curr: u32,
     done: bool,
 }
 
 impl Collatz {
-    pub fn new(start: u128) -> Collatz {
+    pub fn new(start: u32) -> Collatz {
         Collatz {
             curr: start,
             done: false,
@@ -17,9 +17,9 @@ impl Collatz {
 }
 
 impl Iterator for Collatz {
-    type Item = u128;
+    type Item = u32;
 
-    fn next(&mut self) -> Option<u128> {
+    fn next(&mut self) -> Option<u32> {
         if self.done { return None; };
         let result = Some(self.curr);
         self.curr = match self.curr {
@@ -41,7 +41,7 @@ fn help() {
 
 #[derive(Debug)]
 struct Result {
-    start: u128,
+    start: u32,
     len: usize,
     index_max: usize,
     max: usize,
@@ -49,7 +49,7 @@ struct Result {
 
 fn calc_slice(slice: &mut [Result]) {
     for (_, e) in slice.iter_mut().enumerate() {
-        let collatz: Vec<u128> = Collatz::new(e.start as u128).into_iter().collect();
+        let collatz: Vec<u32> = Collatz::new(e.start as u32).into_iter().collect();
         let max_value = collatz.iter().fold(e.start, |max, &val| if val > max { val } else { max });
         let index_of_max = collatz.iter().position(|&r| r == max_value).unwrap();
 
@@ -57,11 +57,11 @@ fn calc_slice(slice: &mut [Result]) {
     }
 }
 
-fn calc(upper_limit: u128) -> Vec<Result> {
+fn calc(upper_limit: u32) -> Vec<Result> {
     let num_cpus = num_cpus::get();
-    let chunk_size: usize = (upper_limit / num_cpus as u128) as usize;
+    let chunk_size: usize = (upper_limit / num_cpus as u32) as usize;
 
-    let mut nums: Vec<u128> = (0..upper_limit).collect();
+    let mut nums: Vec<u32> = (0..upper_limit).collect();
     let mut table: Vec<Result> = nums.iter_mut().map(|i| Result { start: *i, len: 0, index_max: 0, max: 0 }).collect();
 
     let _ = crossbeam::scope(|scope| {
@@ -76,11 +76,12 @@ fn calc(upper_limit: u128) -> Vec<Result> {
 }
 
 fn main() {
+
     let args: Vec<String> = env::args().collect();
     match args.len() {
         2 => {
             println!("{:?}", args);
-            match args[1].parse::<u128>() {
+            match args[1].parse::<u32>() {
                 Ok(upper_limit) => {
                     println!("Calculating ...");
                     for record in calc(upper_limit +1 ) {
